move midi listener
